requirements.txt
virtualenv
requests
sqlalchemy
selenium
mimesis
pymysql
pytest
pytest-allure
coloredlogs
pyyaml
ruamel.yaml
click
requests_toolbelt
redis